Move trigger tests to new test file
