+++ lib/dialects/mssql/query/compiler.js
@@ -185,14 +185,14 @@ class QueryCompiler_MSSQL extends QueryCompiler {
   }
 
   _buildReturningDeclare(values) {
-    if (values) {
+    if (values && values.length > 0) {
       const columns = returning.map(value => `[${value}] nvarchar(max) null`);
       return `declare @out table (${columns.join(",")}); `;
     }
   }
 
   _buildReturningSelect(values) {
-    if (values) {
+    if (values && values.length > 0) {
       const selections = returning.map(value => `[${value}]`);
       return `; select ${selections.join(",")} from @out`
     }
