+++ test/integration/builder/inserts.js
@@ -1817,81 +1817,5 @@ module.exports = function (knex) {
           });
       });
     });
-
-    it('#4152 Should allow returns with inserts on tables with triggers', async function () {
-      if (knex.client.driverName !== 'mssql') {
-        return true;
-      }
-
-      return await knex.transaction(async function () {
-        // Reused variables
-        const primaryTable = 'test_return_with_trigger_primary';
-        const secondaryTable = 'test_return_with_trigger_secondary';
-        const primaryLink = 'foreign_id';
-        const secondaryLink = 'looping_id';
-        const triggerName = 'tr_test_return_with_trigger';
-        let insertResults;
-
-        async function setup() {
-          // Create tables
-          await knex.schema.createTable(primaryTable, function (table) {
-            table.increments();
-            table.string("data");
-            table.integer(primaryLink);
-          });
-
-          await knex.schema.createTable(secondaryTable, function (table) {
-            table.increments();
-            table.string("data");
-            table.integer(secondaryLink);
-          });
-
-          await knex.raw(`
-            GO
-            /****** Object:  Trigger [dbo].[${triggerName}]    Script Date: 12/9/2020 10:52:56 AM ******/
-            SET ANSI_NULLS ON
-            GO
-            SET QUOTED_IDENTIFIER ON
-            GO
-            CREATE TRIGGER [dbo].[${triggerName}] ON [dbo].[${secondaryTable}]
-              AFTER INSERT
-            AS 
-            BEGIN
-              -- SET NOCOUNT ON added to prevent extra result sets from
-              -- interfering with SELECT statements.
-              SET NOCOUNT ON;
-
-              BEGIN
-                update pt
-                set pt.${primaryLink} = i.id
-                from Inserted as i
-                  inner join dbo.${primaryTable} as pt
-                    on pt.id = i.${secondaryLink}
-              END
-            END
-          `);
-        }
-
-        async function insertWithReturn() {
-          const primaryId = await knex(primaryTable).insert([insertPrimary], ["id"]);
-          insertSecondary[secondaryLink] = primaryId[0];
-
-          // Test retrieve with trigger
-          insertResults = await knex(secondaryTable).insert([insertSecondary], ["id"]);
-        }
-
-        async function cleanup() {
-          await knex.raw(`drop trigger ${triggerName}`);
-          await knex.schema.dropTable(primaryTable);
-          await knex.schema.dropTable(secondaryTable);
-        }
-
-        await setup();
-        await insertWithReturn();
-        await cleanup();
-
-        expect(insertResults).to.be.finite;
-      });
-    });
   });
 };
